* 'master' of gitlab.ayla.com.cn:ccpg-hotel/construction-android:
  hotfix-SAAS-950-maxLine=1-导致文本丢失问题修改。
  1.场景列表功能匹配的问题修改。
  1.程序icon修改。
  1.cd规则修改。
  1.CD规则调整：所有分支触发lint、master，prod触发assembleRelease、prod触发publish。 2.icon修改。

+++ app/src/main/java/com/ayla/hotelsaas/bean/DeviceCategoryDetailBean.java
@@ -17,7 +17,7 @@ public class DeviceCategoryDetailBean {
      */
 
     private String deviceName;
-    private long cuId;
+    private int cuId;
     private String oemModel;
     private List<String> conditionProperties;
     private List<String> actionProperties;
@@ -30,11 +30,11 @@ public class DeviceCategoryDetailBean {
         this.deviceName = deviceName;
     }
 
-    public long getCuId() {
+    public int getCuId() {
         return cuId;
     }
 
-    public void setCuId(long cuId) {
+    public void setCuId(int cuId) {
         this.cuId = cuId;
     }
